Update team section with new member and details
This commit updates the team section in `index.html` by revising the description for an existing team member, Мария, to better reflect her experience with "5 years leading marketing initiatives" instead of the previous "UAM, Growth Hacking". Additionally, a new team member, Даниил Вихров, specializing in PR and public relations, has been added. His profile includes his role as a Project Manager and PR Director, along with his experience working with notable organizations such as РЖД, ArtMasters, and World of Dogs. This update aims to provide a more accurate and enriched presentation of the team's composition and expertise.

A minor change was also made to `foto.png`, but this commit primarily focuses on enhancing the team's representation on the website.