Add i18n support and update team member info
This commit introduces internationalization support for team member
details in `index.html` by adding `data-i18n` attributes. It ensures
that the website can be easily translated into different languages
without altering the structure of the HTML.

Additionally, the commit updates the `main.js` file to include a new
team member, Daniel Vikhrov, under the "Public Relations and Community
Engagement" position. It also corrects the position title of Sevak
Petrosyan to "СТО" to maintain consistency with the newly introduced
internationalization approach.

These changes are part of an ongoing effort to make the website more
accessible to a global audience and to keep the team information up to
date.